Remove Azure deployment workflow for Node.js project
This commit deletes the GitHub Actions YAML file used for building and deploying a Node.js project to an Azure Function App. The file was located in the `.github/workflows` directory and named `feature-function-login_postech-fiap-serverless.yml`. This change may indicate a shift in deployment strategy or technology used for this particular project.